Fix onDragEvent validation in FileUpload
Workaround for empty mimetypes issue.
[react-dropzone issue#276](https://github.com/react-dropzone/react-dropzone/issues/276#issuecomment-308637324)

Getting mimetype in onDragEvent was behaving differently
on different browsers and platforms.
(i.e.) When the Google Earth Pro application is not installed.
In google chrome, mimetypes of .kml and .kmz files return empty.
However, it has been observed that "mimetypes"
are returned correctly in requests made with Postman.

As a result, empty mimetypes were added to accepted types and
implement extra validation on empty mimetypes .